+++ src/utils/resolvers.ts
@@ -1,4 +1,4 @@
-import { SYMBOL_OBSERVABLE_READABLE, SYMBOL_UNCACHED, SYMBOL_OBSERVABLE_WRITABLE, SYMBOL_DOM } from '../constants'
+import { SYMBOL_OBSERVABLE_READABLE, SYMBOL_UNCACHED, SYMBOL_OBSERVABLE_WRITABLE, /* SYMBOL_DOM */ } from '../constants'
 import { isObservable } from '../methods/soby'
 import { useRenderEffect } from '../hooks/use_render_effect'
 import { $$ } from '../methods/soby'
@@ -7,23 +7,23 @@ import { isArray, isFunction, isFunctionReactive, isString } from '../utils/lang
 import type { Classes, ObservableMaybe, Styles } from '../types'
 import { Observable, Stack } from '../soby'
 
-const replaceSelf = <T extends { [SYMBOL_DOM]: HTMLElement | HTMLElement[] } & Observable<HTMLElement>>(value: T, newNode: HTMLElement | HTMLElement[]) => {
-  const node = value[SYMBOL_DOM]
-  if (!node)
-    return false
+// const replaceSelf = <T extends { [SYMBOL_DOM]: HTMLElement | HTMLElement[] } & Observable<HTMLElement>>(value: T, newNode: HTMLElement | HTMLElement[]) => {
+//   const node = value[SYMBOL_DOM]
+//   if (!node)
+//     return false
 
-  const isList = newNode instanceof NodeList
+//   const isList = newNode instanceof NodeList
 
-  if (node instanceof NodeList || isArray(node)) {
-    const ns = [...(node as any)].flat()
-    ns.forEach((n, i) => i !== 0 && n.remove())
-    ns[0].replaceWith(...(value[SYMBOL_DOM] = (isList ? [...newNode as any] : [newNode]).flat()))
-  }
-  else
-    node.replaceWith(...(value[SYMBOL_DOM] = (isList ? [...newNode as any] : [newNode]).flat()))
+//   if (node instanceof NodeList || isArray(node)) {
+//     const ns = [...(node as any)].flat()
+//     ns.forEach((n, i) => i !== 0 && n.remove())
+//     ns[0].replaceWith(...(value[SYMBOL_DOM] = (isList ? [...newNode as any] : [newNode]).flat()))
+//   }
+//   else
+//     node.replaceWith(...(value[SYMBOL_DOM] = (isList ? [...newNode as any] : [newNode]).flat()))
 
-  return true
-}
+//   return true
+// }
 
 export const resolveChild = <T>(value: ObservableMaybe<T>, setter: ((value: T | T[], dynamic: boolean, stack: Stack) => void), _dynamic: boolean = false, stack: Stack): void => {