+++ src/utils/setters.ssr.ts
@@ -2,10 +2,10 @@ import { DIRECTIVES, SYMBOLS_DIRECTIVES, SYMBOL_UNCACHED } from '../constants'
 import { useMicrotask } from '../hooks/use_microtask'
 import { useRenderEffect } from '../hooks/use_render_effect'
 import { isStore } from '../methods/soby'
-import { $$ } from '../methods/soby'
+import { $$, isObservable } from '../methods/soby'
 import { store } from '../methods/soby'
 import { untrack } from '../methods/soby'
-import { context, with as _with } from 'soby'
+import { context, with as _with, SYMBOL_OBSERVABLE_WRITABLE } from 'soby'
 import { SYMBOL_STORE_OBSERVABLE } from 'soby'
 import { classesToggle } from '../utils/classlist'
 import { createText, createComment } from '../utils/creators.ssr'
@@ -68,13 +68,19 @@ export const setAttributeStatic = (() => {
 export const setAttribute = (element: HTMLElement, key: string, value: FunctionMaybe<null | undefined | boolean | number | string>, stack: Stack): void => {
 
     if (isFunction(value) && isFunctionReactive(value)) {
-
-        useRenderEffect(() => {
-
-            setAttributeStatic(element, key, value())
-
-        }, stack)
-
+        // Check if value is an observable with toHtml option
+        if (isObservable(value) && value[SYMBOL_OBSERVABLE_WRITABLE]?.options?.toHtml) {
+            useRenderEffect(() => {
+                const unwrappedValue = value()
+                const options = value[SYMBOL_OBSERVABLE_WRITABLE].options
+                const htmlValue = options.toHtml(unwrappedValue)
+                setAttributeStatic(element, key, htmlValue)
+            }, stack)
+        } else {
+            useRenderEffect(() => {
+                setAttributeStatic(element, key, value())
+            }, stack)
+        }
     } else {
 
         setAttributeStatic(element, key, $$(value))